Added functions for signInActivity
Added a functions that will check if the sign in details are valid.
Added a progressbar that will show up when loading after pressing on sign in button.
Added a sign in function that will check the stored data at Firebase if the user
trying to login is an registered user.