Added a listenMessages function.
listenMessages will listen for new chat messages in the database.
Whenever a new message is sent, the EventListener is called that will
add the message to a sorted list. The ChatAdapter is then used with
RecyclerView to display the message for each user in the ChatActivity
window.

Problem: The sender of a chat message will only see a date in their own
chat bubbles, while the reciever will see the senders message just fine.
Will troubleshoot and debug at a later time.